added notification

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -50,7 +50,7 @@ export async function GET(
 
 export async function POST(
   req: Request,
-  { params: { postId } }: { params: { postId: string } },
+  { params  }: { params: { postId: string } },
 ) {
   try {
     const session = await auth();
@@ -58,20 +58,45 @@ export async function POST(
     if (!session || !session.user || !session.user.id) {
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
+    const {postId}=await params;
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
+      },
+    });
 
-    await prisma.like.upsert({
-      where: {
-        userId_postId: {
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where: {
+          userId_postId: {
+            userId: session.user.id,
+            postId,
+          },
+        },
+        create: {
           userId: session.user.id,
           postId,
         },
-      },
-      create: {
-        userId:  session.user.id,
-        postId,
-      },
-      update: {},
-    });
+        update: {},
+      }),
+      ...(session.user.id !== post.userId
+        ? [
+            prisma.notification.create({
+              data: {
+                issuerId: session.user.id,
+                recipientId: post.userId,
+                postId,
+                type: "LIKE",
+              },
+            }),
+          ]
+        : []),
+    ]);
 
     return new Response();
   } catch (error) {
@@ -91,12 +116,33 @@ export async function DELETE(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    await prisma.like.deleteMany({
-      where: {
-        userId:  session.user.id,
-        postId,
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
       },
     });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: {
+          userId: session.user.id,
+          postId,
+        },
+      }),
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: session.user.id,
+          recipientId: post.userId,
+          postId,
+          type: "LIKE",
+        },
+      }),
+    ]);
 
     return new Response();
   } catch (error) {
